herptracker/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.3
requests==2.28.1